FixBugs

+++ backend/app/models/reseller.py
@@ -4,7 +4,7 @@ import enum
 
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, Enum, ForeignKey, Integer, JSON, String, Text
+from sqlalchemy import BigInteger, Boolean, DateTime, Enum, ForeignKey, Integer, JSON, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.db import Base
@@ -36,7 +36,8 @@ class Reseller(Base, TimestampMixin):
     can_create_subreseller: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False)
     status: Mapped[ResellerStatus] = mapped_column(Enum(ResellerStatus), default=ResellerStatus.active, nullable=False)
 
-    balance: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
+    # BigInteger so high-value (e.g. Rial-denominated) wallets cannot overflow int32.
+    balance: Mapped[int] = mapped_column(BigInteger, default=0, nullable=False)
 
     # default pricing for this reseller
     price_per_gb: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
